mnv2_first: rename param_store.py -> store.py
